--- packages/mds-compliance-service/service/repositories/compliance-response/entities/index.ts
@@ -1,1 +1,0 @@
-export { ComplianceResponsePersistence } from './compliance-response-entity'
